+++ tiger-admin/src/main/java/de/gematik/test/tiger/admin/TigerAdminApplication.java
@@ -1,0 +1,12 @@
+package de.gematik.test.tiger.admin;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class TigerAdminApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(TigerAdminApplication.class, args);
+    }
+}