+++ tiger-proxy/src/main/java/de/gematik/test/tiger/proxy/AbstractTigerProxy.java
@@ -23,6 +23,7 @@ import de.gematik.rbellogger.data.RbelElement;
 import de.gematik.rbellogger.key.RbelKey;
 import de.gematik.test.tiger.common.config.tigerProxy.TigerProxyConfiguration;
 import de.gematik.test.tiger.common.config.tigerProxy.TigerRoute;
+import de.gematik.test.tiger.exception.TigerProxyStartupException;
 import lombok.Data;
 import lombok.extern.slf4j.Slf4j;
 
@@ -39,6 +40,9 @@ public abstract class AbstractTigerProxy implements ITigerProxy {
     private RbelLogger rbelLogger;
 
     public AbstractTigerProxy(TigerProxyConfiguration configuration) {
+        if (configuration.getTls() == null) {
+            throw new TigerProxyStartupException("no TLS-configuration found!");
+        }
         rbelLogger = buildRbelLoggerConfiguration(configuration)
             .constructRbelLogger();
         if (configuration.isDisableRbelParsing()) {